src/test/java/com/blogspot/ostas/leetcode/all/medium/maximum_product_subarray/SolutionTest.java
package com.blogspot.ostas.leetcode.all.medium.maximum_product_subarray;

import org.junit.jupiter.api.Test;

import static org.assertj.core.api.Assertions.assertThat;

class SolutionTest {

    /*
      Example 1:
    Input: nums = [2,3,-2,4]
    Output: 6
    Explanation: [2,3] has the largest product 6.
    Example 2:
    Input: nums = [-2,0,-1]
    Output: 0
    Explanation: The result cannot be 2, because [-2,-1] is not a subarray.
      Constraints:
    1 <= nums.length <= 2 * 104
    -10 <= nums[i] <= 10
    The product of any prefix or suffix of nums is guaranteed to fit in a 32-bit integer.
    */
    @Test
    void example_0() {
        var solution = new Solution();
        int[] nums = new int[]{2, 3, -2, 4};
        int expected = 6;
        var result = solution.maxProduct(nums);
        assertThat(result).isEqualTo(expected);
    }

    @Test
    void example_1() {
        var solution = new Solution();
        int[] nums = new int[]{-2, 0, -1};
        int expected = 0;
        var result = solution.maxProduct(nums);
        assertThat(result).isEqualTo(expected);
    }

}
